Use new `RecordCipher` class for record secret encryption.
